feature/cb2-12317: Refactor into services (#187)
* moved test checks to test service

* moved tech record methods to tech record service

* moved defect methods to defect service

* pr review removed defect service function
